rename knob and led component files to match v-piano-keyboard
